Component development. -Page-component superType corrected. -"topnav" changed in order to list inner pages. Page, even inner, are listed in one line.
