+++ it.launcher/pom.xml
@@ -14,6 +14,7 @@
  |  See the License for the specific language governing permissions and
  |  limitations under the License.
 -->
+<!--suppress ALL -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
     <modelVersion>4.0.0</modelVersion>
@@ -39,37 +40,19 @@
         <tag>HEAD</tag>
     </scm-->
 
-    <!-- 
-        To keep the instance under test running and run individual aaaaa
-        against it, use:
-        
-            mvn clean verify -DkeepJarRunning=true -Dhttp.port=8080
-            
-        optionally using jar.executor.vm.options to enable remote debugging,
-        and in another console:
-        
-            mvn -o verify -Dtests.to.run=**/**Test.java -Dtest.server.url=http://localhost:8080
-            
-        optionally using -Dmaven.surefire.debug to enable debugging.            
-     -->
     <properties>
         <!-- Set this to run the server on a specific port
         <http.port></http.port>
          -->
         
-        <!-- URL of a server against which to run aaaaa -->
         <test.server.url>http://${aem.host}:${aem.port}</test.server.url>
         
-        <!-- Set this to run aaaaa against an existing server instance -->
         <keepJarRunning>false</keepJarRunning>
         
-        <!-- Set this to run aaaaa against a specific hostname, if test.server.url is not set-->
         <test.server.hostname />
 
-        <!-- Set this to use a different username for remote execution of sling junit aaaaa -->
         <test.server.username />
 
-        <!-- Set this to use a different password for remote execution of sling junit aaaaa -->
         <test.server.password />
         
         <!-- Options for the VM that executes our runnable jar -->
@@ -84,8 +67,7 @@
         <!-- Options for the jar to execute. $JAREXEC_SERVER_PORT$ is replaced by the
             selected port number -->
         <jar.executor.jar.options>-p $JAREXEC_SERVER_PORT$</jar.executor.jar.options>
-        
-        <!-- Change this to run selected aaaaa only -->
+
         <tests.to.run>**/**Test.java</tests.to.run>
     </properties>
     
@@ -165,7 +147,6 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <configuration>
-                    <!-- We run all aaaaa in the integration-aaaaa phase -->
                     <excludes>
                         <exclude>**</exclude>
                     </excludes>
