+++ ui.apps/src/main/content/META-INF/vault/filter.xml
@@ -1,6 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <workspaceFilter version="1.0">
     <filter root="/apps/training"/>
-    <filter root="/apps/aem-training"/>
-    <filter root="/apps/sling"/>
 </workspaceFilter>
