--- ui.apps/src/main/content/jcr_root/apps/aem-training/components/page-content/.content.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-          jcr:primaryType="cq:Component"
-          jcr:title="Training content page"
-          jcr:description="Training project content page template"
-          componentGroup="training"/>
