--- ui.apps/src/main/content/jcr_root/apps/aem-training/components/page-content/page-content.html
@@ -1,19 +1,0 @@
-<html>
-  <head>
-    <title>Hello World!</title>
-  </head>
-  <body>
-    <h1>Hello world!</h1>
-    <h2>I’m your rendering script!!</h2>
-
-    <div>
-      <div>logo</div>
-      <div>
-        <div>toptoolbar</div>
-      </div>
-      <div data-sly-resource="${'topnav' @ resourceType='training/components/topnav'}"></div>
-    </div>
-
-  </body>
-</html>
-
