--- ui.apps/src/main/content/jcr_root/apps/aem-training/components/topnav/topnav.html
@@ -1,3 +1,0 @@
-<ul class=”topnav” data-sly-use.topnav=”TopNav” data-sly-list=”${topnav.items}”>
-  <li><a href=”${item.path}.html”>${item.title}</a></li>
-</ul>
