--- ui.apps/src/main/content/jcr_root/apps/training/aaaaa/SampleTests.js
@@ -1,28 +1,0 @@
-/*
- *  Copyright 2015 Adobe Systems Incorporated
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-new hobs.TestSuite("training Tests", {path:"/apps/training/tests/SampleTests.js", register: true})
-
-    .addTestCase(new hobs.TestCase("Hello World component on english page")
-        .navigateTo("/content/training/en.html")
-        .asserts.location("/content/training/en.html", true)
-        .asserts.visible(".helloworld", true)
-    )
-
-    .addTestCase(new hobs.TestCase("Hello World component on french page")
-        .navigateTo("/content/training/fr.html")
-        .asserts.location("/content/training/fr.html", true)
-        .asserts.visible(".helloworld", true)
-    );
