--- ui.apps/src/main/content/jcr_root/apps/training/clientlibs/clientlib-base/.content.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:ClientLibraryFolder"
-    allowProxy="{Boolean}true"
-    categories="[training.base]"
-    embed="[core.wcm.components.image.v1,core.wcm.components.breadcrumb.v1,core.wcm.components.form.container.v1,core.wcm.components.form.text.v1,core.wcm.components.list.v1,training.breadcrumb]"/>
