--- ui.apps/src/main/content/jcr_root/apps/training/components/content/breadcrumb/.content.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:Component"
-    jcr:title="Breadcrumb"
-    sling:resourceSuperType="core/wcm/components/breadcrumb/v1/breadcrumb"
-    componentGroup="training"/>
