--- ui.apps/src/main/content/jcr_root/apps/training/components/content/breadcrumb/clientlib/less/breadcrumb.less
@@ -1,30 +1,0 @@
-/*
- *  Copyright 2017 Adobe Systems Incorporated
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-.breadcrumb {
-    display: inline-block;
-    list-style: none;
-    padding: 0;
-
-    & .breadcrumb-item {
-        float: left;
-
-        & + .breadcrumb-item::before {
-            display: inline-block; //
-            padding: 0 10px;
-            content: "\003e";
-        }
-    }
-}
