--- ui.apps/src/main/content/jcr_root/apps/training/components/content/helloworld/.content.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:Component"
-    jcr:title="Hello World Component"
-    componentGroup="training"/>
