--- ui.apps/src/main/content/jcr_root/apps/training/components/content/image/_cq_editConfig.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
-    jcr:primaryType="cq:EditConfig">
-    <cq:dropTargets jcr:primaryType="nt:unstructured">
-        <image
-            jcr:primaryType="cq:DropTargetConfig"
-            accept="[image/.*]"
-            groups="[media]"
-            propertyName="./fileReference">
-            <parameters
-                jcr:primaryType="nt:unstructured"
-                sling:resourceType="training/components/content/image"
-                imageCrop=""
-                imageMap=""
-                imageRotate=""/>
-        </image>
-    </cq:dropTargets>
-    <cq:inplaceEditing
-        jcr:primaryType="cq:InplaceEditingConfig"
-        active="{Boolean}true"
-        editorType="image">
-        <config jcr:primaryType="nt:unstructured">
-            <plugins jcr:primaryType="nt:unstructured">
-                <crop
-                    jcr:primaryType="nt:unstructured"
-                    features="*"/>
-                <flip
-                    jcr:primaryType="nt:unstructured"
-                    features="-"/>
-                <map
-                    jcr:primaryType="nt:unstructured"
-                    features="*"/>
-                <rotate
-                    jcr:primaryType="nt:unstructured"
-                    features="*"/>
-                <zoom
-                    jcr:primaryType="nt:unstructured"
-                    features="*"/>
-            </plugins>
-            <ui jcr:primaryType="nt:unstructured">
-                <inline
-                    jcr:primaryType="nt:unstructured"
-                    toolbar="[crop#launch,rotate#right,history#undo,history#redo,fullscreen#fullscreen,control#close,control#finish]">
-                    <replacementToolbars
-                        jcr:primaryType="nt:unstructured"
-                        crop="[crop#identifier,crop#unlaunch,crop#confirm]"/>
-                </inline>
-                <fullscreen jcr:primaryType="nt:unstructured">
-                    <toolbar
-                        jcr:primaryType="nt:unstructured"
-                        left="[crop#launchwithratio,rotate#right,flip#horizontal,flip#vertical,zoom#reset100,zoom#popupslider]"
-                        right="[history#undo,history#redo,fullscreen#fullscreenexit]"/>
-                    <replacementToolbars jcr:primaryType="nt:unstructured">
-                        <crop
-                            jcr:primaryType="nt:unstructured"
-                            left="[crop#identifier]"
-                            right="[crop#unlaunch,crop#confirm]"/>
-                        <map
-                            jcr:primaryType="nt:unstructured"
-                            left="[map#rectangle,map#circle,map#polygon]"
-                            right="[map#unlaunch,map#confirm]"/>
-                    </replacementToolbars>
-                </fullscreen>
-            </ui>
-        </config>
-    </cq:inplaceEditing>
-</jcr:root>
