--- ui.apps/src/main/content/jcr_root/apps/training/components/content/list/_cq_editConfig.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
-    jcr:primaryType="cq:EditConfig">
-    <cq:dropTargets jcr:primaryType="cq:DropTargetConfig">
-        <pages
-            jcr:primaryType="nt:unstructured"
-            accept="[*]"
-            groups="[page]"
-            propertyName="./pages">
-            <parameters
-                jcr:primaryType="nt:unstructured"
-                listFrom="static"/>
-        </pages>
-    </cq:dropTargets>
-</jcr:root>
