--- ui.apps/src/main/content/jcr_root/apps/training/components/form/container/.content.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:Component"
-    jcr:title="Form Container"
-    jcr:description="Container for components that allow users to submit information"
-    sling:resourceSuperType="core/wcm/components/form/container/v1/container"
-    componentGroup="training Form"
-    cq:isContainer="{Boolean}true"/>
