+++ ui.apps/src/main/content/jcr_root/apps/training/components/page-content/.content.xml
@@ -3,5 +3,5 @@
           jcr:primaryType="cq:Component"
           jcr:title="Training content page"
           jcr:description="Training project content page template"
-          sling:resourceSuperType="/training/templates/page-content/"
+          sling:resourceSuperType="wcm/foundation/components/page"
           componentGroup="training"/>
