+++ ui.apps/src/main/content/jcr_root/apps/training/components/page-content/page-content.html
@@ -1,3 +1,9 @@
+<html>
+<head>
+    <title>Hello World!</title>
+</head>
+<body>
+
 <div>
     <div>logo</div>
     <div>
@@ -5,3 +11,8 @@
     </div>
     <div data-sly-resource="${'topnav' @ resourceType='training/components/topnav'}"></div>
 </div>
+
+<h1>Hello world!</h1>
+<h2>I’m your rendering script!!</h2>
+</body>
+</html>
