--- ui.apps/src/main/content/jcr_root/apps/training/config.author/com.day.cq.wcm.mobile.core.impl.MobileEmulatorProvider-training.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="sling:OsgiConfig"
-    mobile.resourceTypes="[training/components/structure/page]"
-    README="Indicate which page resource types should display the mobile emulators."/>
