--- ui.apps/src/main/content/jcr_root/apps/training/config/org.apache.sling.commons.log.LogManager.factory.config-training.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0"
-    xmlns:jcr="http://www.jcp.org/jcr/1.0" jcr:primaryType="sling:OsgiConfig"
-    org.apache.sling.commons.log.file="logs/project-training.log"
-    org.apache.sling.commons.log.level="info"
-    org.apache.sling.commons.log.names="[training-project]"
-    org.apache.sling.commons.log.pattern="\{0,date,yyyy-MM-dd HH:mm:ss.SSS} {4} [{3}] {5}" />
