--- ui.apps/src/main/content/jcr_root/apps/training/i18n/fr.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:mix="http://www.jcp.org/jcr/mix/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
-    jcr:language="fr"
-    jcr:mixinTypes="[mix:language]"
-    jcr:primaryType="nt:folder">
-    <copyright
-        jcr:mixinTypes="[sling:Message]"
-        jcr:primaryType="nt:folder"
-        sling:key="&amp;copy; {0} training Site. All rights reserved."
-        sling:message="&amp;copy; {0} training Site. Tous droits réservés."/>
-</jcr:root>
