+++ ui.apps/src/main/content/jcr_root/apps/training/templates/page-content/.content.xml
@@ -2,29 +2,11 @@
 <jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
     jcr:description="Template with focus on content."
     jcr:primaryType="cq:Template"
-    jcr:title="training Content Page"
-    allowedPaths="[/content/training/.+]"
-    ranking="{Long}100">
+    jcr:title="Training Content Page"
+    allowedPaths="/content(/.*)?"
+    ranking="{Long}1">
     <jcr:content
         jcr:primaryType="cq:PageContent"
-        sling:resourceType="training/components/structure/page">
-        <logo
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="training/components/structure/logo"/>
-        <title
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="training/components/content/title"/>
-        <topnav
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="training/components/structure/topnav"/>
-        <par
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="wcm/foundation/components/parsys">
-            <text
-                jcr:primaryType="nt:unstructured"
-                sling:resourceType="training/components/content/text"
-                text="&lt;p>This is a content page&lt;/p>"
-                textIsRich="true"/>
-        </par>
+        sling:resourceType="training/components/page-content">
     </jcr:content>
 </jcr:root>
