--- ui.apps/src/main/content/jcr_root/apps/training/templates/page-home/.content.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:description="Template for initial or main page of the site."
-    jcr:primaryType="cq:Template"
-    jcr:title="training Home Page"
-    allowedPaths="[/content/training]"
-    ranking="{Long}100">
-    <jcr:content
-        jcr:primaryType="cq:PageContent"
-        sling:resourceType="training/components/structure/page">
-        <logo
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="training/components/structure/logo"/>
-        <title
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="training/components/content/title"/>
-        <topnav
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="training/components/structure/topnav"/>
-        <par
-            jcr:primaryType="nt:unstructured"
-            sling:resourceType="wcm/foundation/components/parsys">
-            <text
-                jcr:primaryType="nt:unstructured"
-                sling:resourceType="training/components/content/text"
-                text="&lt;p>This is a home page&lt;/p>"
-                textIsRich="true"/>
-        </par>
-    </jcr:content>
-</jcr:root>
