--- ui.apps/src/main/content/jcr_root/apps/training/tests/.content.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-    jcr:primaryType="cq:ClientLibraryFolder"
-    categories="[cq.authoring.editor.hook.tests]"
-    dependencies="[cq.authoring.editor]"/>
