--- ui.apps/src/main/content/jcr_root/apps/training/tests/js.txt
@@ -1,1 +1,0 @@
-SampleTests.js
